chore(fw): cargo update + fix clippy warnings + fmt
- cargo update: sntpc 0.10→0.11, sntpc-net-embassy 0.11, sntpc-time-embassy
  0.6, plus routine bumps (bitflags, syn, jiff, ...).
- .cargo/config.toml: suppress `invalid_runtime_symbol_definitions` — the new
  deny-by-default nightly lint fires on vendored esp-alloc's `extern memcpy`
  decl only because esp-hal is a [patch] path dep (git deps get --cap-lints
  allow). Keeps the vendored fork == upstream/main untouched.
- clippy: drop redundant `as i64` on sntpc 0.11 `offset()` (now returns i64);
  remove double-parens in mk_static! macro.
- cargo fmt.

+++ meteo/src/ntp_client.rs
@@ -95,7 +95,7 @@ async fn try_sync(stack: Stack<'_>) -> bool {
 
     let Some(pp) = res else { return false };
 
-    let off = pp.offset() as i64;
+    let off = pp.offset();
     unsafe { CURRENT_OFFSET.lock_mut(|s| *s = off) };
     true
 }
